Cleaned up files using partial classes
